Add Quick events support in Quickspace
This adds a new UI to Quickspace when events are available such as Now
Playing, Alarms, or whatever we configure as an event. This logic is
inspired by Google's event infrastructure in Smartspace but completely
written from scratch to be housed in Quickspace.

Initial events are:

- Device intro: Shows on first boot and setup
- Ambient Play (Now Playing): Shows what song is playing and tapping
the song will launch the results in Google search

spezi77 edits:
- Fix package name for DoubleShadowTextView
- Update strings for ROM intro
- Update weather client package

Change-Id: Ia9c5335d77d3ce14d71d133bebf1d8cb509b0318
Signed-off-by: spezi77 <spezi7713@gmx.net>